requirements.txt
autopep8==1.5
#gunicorn==20.0.4
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
account<=0.1.0
#Django==3.0.3
# The user requested celery<=5.2.7
# account 0.1.0 depends on celery<5.0.0 and >=4.4.2
accounts<=0.2.4
# The user requested pytest<=7.1.3
#  accounts 0.2.4 depends on pytest<6.0.0 and >=5.4.1
alembic<=1.8.1
amqp<=5.1.1
anyio<=3.6.1
asgiref<=3.5.2
async-generator<=1.10
atomicwrites<=1.4.1
attrs<=22.1.0
bcrypt<=4.0.0
billiard<=3.6.4.0
cachetools<=5.2.0
celery<=5.2.7
certifi<=2022.6.15.1
cffi<=1.15.1
chardet<=5.0.0
charset-normalizer<=2.0.12
click<=8.1.3
click-didyoumean<=0.3.0
click-plugins<=1.1.1
click-repl<=0.2.0
colorama<=0.4.5
convertdate<=2.4.0
cryptography<=38.0.1
cssselect<=1.1.0
cssutils<=2.6.0
datefinder<=0.7.3
dj-database-url<=1.0.0
#docker
#Django<=4.1.1
Django>=3.0,<4.0
psycopg2>=2.8
#docker
django-crispy-forms<=1.14.0
django-cron<=0.6.0
django-crontab<=0.7.1
django-extensions<=3.2.1
django-filemaker-pyodbc<=0.0.13
django-formtools<=2.3
django-lockdown<=4.0.0
django-otp<=1.1.3
django-phonenumber-field<=6.4.0
django-tempus-dominus<=5.1.2.17
django-two-factor-auth<=1.14.0
dnspython<=2.2.1
docopt<=0.6.2
ecdsa<=0.18.0
email-to<=0.1.0
email-validator<=1.2.1
emails<=6.0
et-xmlfile<=1.1.0
fastapi<=0.83.0
google-api-core<=2.10.0
google-api-python-client<=2.60.0
google-auth<=2.11.0
google-auth-httplib2<=0.1.0
google-auth-oauthlib<=0.5.2
googleapis-common-protos<=1.56.4
greenlet<=1.1.3
gunicorn==20.1.0
h11<=0.13.0
hijri-converter<=2.2.4
holidays<=0.15
httplib2<=0.20.4
idna<=3.3
include<=0.2.2
inflect<=6.0.0
inflect<=5.3.0

iniconfig<=1.1.1
Jinja2<=3.1.2
keyboard<=0.13.5
kombu<=5.2.4
korean-lunar-calendar<=0.2.1
lxml<=4.9.1
Mako<=1.2.2
manage.py<=0.2.10
Markdown<=3.4.1
MarkupSafe<=2.1.1
more-itertools<=8.14.0
MouseInfo<=0.1.3
mysql-connector<=2.2.9
mysql-connector-python<=8.0.30
mysqlclient<=2.1.1
numpy<=1.23.3
oauthlib<=3.2.1
opencv-python<=4.6.0.66
openpyxl<=3.0.10
outcome<=1.2.0
packaging<=21.3
pandas<=1.4.4
passlib<=1.7.4
phonenumbers<=8.12.55
Pillow<=9.2.0
pipreqs<=0.4.11
pluggy<=1.0.0
premailer<=3.10.0
prompt-toolkit<=3.0.31
protobuf<=4.21.5
psycopg2-binary<=2.9.3
py<=1.11.0
pyasn1<=0.4.8
pyasn1-modules<=0.2.8
PyAutoGUI<=0.9.53
pycparser<=2.21
pydantic<=1.10.2
PyGetWindow<=0.0.9
PyMeeus<=0.5.11
PyMsgBox<=1.0.9
pynput<=1.7.6
pyOpenSSL<=22.0.0
pyparsing<=3.0.9
PyPDF2<=2.10.8
pyperclip<=1.8.2
PyRect<=0.2.0
PyScreeze<=0.1.28
PySocks<=1.7.1
pytest<=7.1.3
python-crontab<=2.6.0
python-dateutil<=2.8.2
python-dotenv<=0.21.0
python-jose<=3.3.0
python-multipart<=0.0.5
pytweening<=1.0.4
pytz<=2022.2.1
#pywin32<=304
qrcode<=7.3.1
raven<=6.10.0
regex<=2022.9.13
reportlab<=3.6.11
requests<=2.28.1
requests-oauthlib<=1.3.1
rsa<=4.9
schedule<=1.1.0
selenium<=4.4.3
six<=1.16.0
sniffio<=1.3.0
sortedcontainers<=2.4.0
SQLAlchemy<=1.4.41
sqlparse<=0.4.2
starlette<=0.20.4
tenacity<=8.0.1
tomli<=2.0.1
tqdm<=4.64.1
trio<=0.21.0
trio-websocket<=0.9.2
typing_extensions<=4.3.0
tzdata<=2022.2
uritemplate<=4.1.1
urllib3<=1.26.12
uvicorn<=0.18.3
vine<=5.0.0
wcwidth<=0.2.5
webdriver-manager<=3.8.3
websockets<=10.3
Werkzeug==2.2.2
whitenoise<=6.2.0
wsproto<=1.2.0
yarg<=0.1.9